requirements.txt
pymongo
python-dotenv
langchain
openai
chainlit
